fix: 🙅 Disallow reused UIDs (#212)
Unfortunately a number of UIDs were reused, especially for things like
automatic updaters. This change introduces the concept of blocked
identifiers to the library configuration file and stops them from being
used for grouping.

The set of identifiers introduced to 'full.yaml' as part of this change
are all updaters (identified by searching for the string 'kB') and
'explodes' 295 EPOC32 programs. 💥